🐛Fix some bugs
